New User Login (setup)
DB - added two fields for web users and new user with default values where applicable. Models updated with new fields;
Front End - Updated detail with flags and javascript to auto-populate username with the entered email address.
Need to just add a check to the new user flag upon login to flip the flag to false when logging in the first time and then redirect to change password page.