Dashboard: Completed
Resized and aligned each section/card to fit the width of the screen. Updated the colors of the chart to differentiate between current vs prev years. Added a new view "SchedServiceKPI", updated SFADBContext, created model, and updated related dashboard files to visually show the data on the dashboard.

Added a link in the upper right hand corner which takes users to the knowledge base.